requirements.txt
# MultispectralResNet 项目依赖
# 基于实际使用的包版本，在 AutoDL RTX 3090 环境测试通过

# 深度学习框架
torch==2.0.0+cu118
torchvision==0.15.1+cu118

# 数值计算和数据处理
numpy==1.24.2
scipy==1.10.1
pandas==2.0.3
scikit-learn==1.3.2

# 图像处理
opencv-python==4.12.0.88
Pillow==9.4.0

# 数据文件处理
openpyxl==3.1.5          # Excel文件读取 

# 可视化
matplotlib==3.7.1

# 工具库
tqdm==4.67.1             # 进度条显示
PyYAML>=6.0              # YAML配置文件解析


